Renames ld-worker to fetch-worker and move all fetches of owner-api into the worker
